// src/enums/datetime.enums.ts
export enum DayOfWeekEnum {
  SUNDAY = 0,
  MONDAY = 1,
  TUESDAY = 2,
  WEDNESDAY = 3,
  THURSDAY = 4,
  FRIDAY = 5,
  SATURDAY = 6,
}

export enum DayOfMonthEnum {
  DAY_1 = 1,
  DAY_2 = 2,
  DAY_3 = 3,
  DAY_4 = 4,
  DAY_5 = 5,
  DAY_6 = 6,
  DAY_7 = 7,
  DAY_8 = 8,
  DAY_9 = 9,
  DAY_10 = 10,
  DAY_11 = 11,
  DAY_12 = 12,
  DAY_13 = 13,
  DAY_14 = 14,
  DAY_15 = 15,
  DAY_16 = 16,
  DAY_17 = 17,
  DAY_18 = 18,
  DAY_19 = 19,
  DAY_20 = 20,
  DAY_21 = 21,
  DAY_22 = 22,
  DAY_23 = 23,
  DAY_24 = 24,
  DAY_25 = 25,
  DAY_26 = 26,
  DAY_27 = 27,
  DAY_28 = 28,
  DAY_29 = 29,
  DAY_30 = 30,
  DAY_31 = 31,
}

export enum HoursEnum {
  HOUR_0 = '00',
  HOUR_1 = '01',
  HOUR_2 = '02',
  HOUR_3 = '03',
  HOUR_4 = '04',
  HOUR_5 = '05',
  HOUR_6 = '06',
  HOUR_7 = '07',
  HOUR_8 = '08',
  HOUR_9 = '09',
  HOUR_10 = '10',
  HOUR_11 = '11',
  HOUR_12 = '12',
  HOUR_13 = '13',
  HOUR_14 = '14',
  HOUR_15 = '15',
  HOUR_16 = '16',
  HOUR_17 = '17',
  HOUR_18 = '18',
  HOUR_19 = '19',
  HOUR_20 = '20',
  HOUR_21 = '21',
  HOUR_22 = '22',
  HOUR_23 = '23',
}

export enum MinutesEnum {
  MINUTE_0 = '00',
  MINUTE_15 = '15',
  MINUTE_30 = '30',
  MINUTE_45 = '45',
}

export enum TimeFormatEnum {
  AM = 'am',
  PM = 'pm',
}